cli: split up `commands/operation.rs` into one file per command
