+++ cli/src/commands/operation/restore.rs
@@ -1,0 +1,59 @@
+// Copyright 2020-2023 The Jujutsu Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use jj_lib::object_id::ObjectId;
+
+use super::{view_with_desired_portions_restored, UndoWhatToRestore, DEFAULT_UNDO_WHAT};
+use crate::cli_util::CommandHelper;
+use crate::command_error::CommandError;
+use crate::ui::Ui;
+
+/// Create a new operation that restores the repo to an earlier state
+///
+/// This restores the repo to the state at the specified operation, effectively
+/// undoing all later operations. It does so by creating a new operation.
+#[derive(clap::Args, Clone, Debug)]
+pub struct OperationRestoreArgs {
+    /// The operation to restore to
+    ///
+    /// Use `jj op log` to find an operation to restore to. Use e.g. `jj
+    /// --at-op=<operation ID> log` before restoring to an operation to see the
+    /// state of the repo at that operation.
+    operation: String,
+
+    /// What portions of the local state to restore (can be repeated)
+    ///
+    /// This option is EXPERIMENTAL.
+    #[arg(long, value_enum, default_values_t = DEFAULT_UNDO_WHAT)]
+    what: Vec<UndoWhatToRestore>,
+}
+
+pub fn cmd_op_restore(
+    ui: &mut Ui,
+    command: &CommandHelper,
+    args: &OperationRestoreArgs,
+) -> Result<(), CommandError> {
+    let mut workspace_command = command.workspace_helper(ui)?;
+    let target_op = workspace_command.resolve_single_op(&args.operation)?;
+    let mut tx = workspace_command.start_transaction();
+    let new_view = view_with_desired_portions_restored(
+        target_op.view()?.store_view(),
+        tx.base_repo().view().store_view(),
+        &args.what,
+    );
+    tx.mut_repo().set_view(new_view);
+    tx.finish(ui, format!("restore to operation {}", target_op.id().hex()))?;
+
+    Ok(())
+}
